+++ WitnessPuzzles/Document/Hexagon.swift
@@ -16,7 +16,7 @@ extension WitnessPuzzlesDocument {
         
         func extent( puzzle: WitnessPuzzlesDocument ) -> CGRect {
             let center = position.puzzle2user( puzzle: puzzle )
-            let radius = CGFloat( puzzle.lineWidth ) / 2
+            let radius = puzzle.lineWidth / 2
             return CGRect(
                 x: center.x - radius, y: center.y - radius, width: 2 * radius, height: 2 * radius
             )
