+++ WitnessPuzzles/Document/Missing.swift
@@ -15,8 +15,8 @@ extension WitnessPuzzlesDocument {
         
         func extent( puzzle: WitnessPuzzlesDocument ) -> CGRect {
             let center = position.puzzle2user( puzzle: puzzle )
-            let short = CGFloat( puzzle.lineWidth )
-            let long = CGFloat( puzzle.lineWidth + puzzle.blockWidth )
+            let short = puzzle.lineWidth
+            let long = puzzle.lineWidth + puzzle.blockWidth
             let width = position.isVertical ? short : long
             let height = position.isVertical ? long : short
